+++ Assets/Scripts/Control/Movement/AStarNode.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 9a03fe6171fe4f19bd110d08c3486131
+timeCreated: 1781190841
