+++ Assets/Scripts/Control/Movement/NPCMoveFocus.cs
@@ -1,0 +1,12 @@
+namespace Frankie.Control
+{
+    public enum NPCMoveFocus
+    {
+        Inactive,
+        Chasing,
+        Fleeing,
+        Pending,
+        Patrolling,
+        RandomWalk
+    }
+}
