+++ Assets/Scripts/Control/Movement/NPCMoveFocus.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b7c07fc55e5a4a48a6f93b6457ed5e74
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
