+++ Assets/Scripts/Control/Movement/NPCMover.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ae64d0078a8ee6e4d8bece5ce559d360
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
