+++ Assets/Scripts/Control/Movement/Patrol/PatrolPathWaypoint.cs
@@ -1,0 +1,13 @@
+using UnityEngine;
+
+namespace Frankie.Control
+{
+    public class PatrolPathWaypoint : MonoBehaviour
+    {
+        // Tunables
+        [SerializeField] WaypointType waypointType = WaypointType.Move;
+
+        // Public Methods
+        public WaypointType GetWaypointType() => waypointType;
+    }
+}
