+++ Assets/Scripts/Control/Movement/Patrol/PatrolPathWaypoint.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 15f3ef01d0185dc4a809bf2c340c2c38
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
