+++ Assets/Scripts/Control/Movement/Patrol/WaypointType.cs
@@ -1,0 +1,8 @@
+namespace Frankie.Control
+{
+    public enum WaypointType
+    {
+        Move,
+        Warp
+    }
+}
