+++ Assets/Scripts/Control/Movement/Patrol/WaypointType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 02c2a09e44f9b7a49897a1b1573e3f19
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
