+++ Assets/Scripts/Control/Movement/PlayerMover.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cfb66abb30066404e974cd0255d37842
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
