--- Assets/Scripts/Control/Mover.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7a853a3a60cfed642953d559e1f8190f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
