--- Assets/Scripts/Control/NPC/NPCMoveFocus.cs
@@ -1,12 +1,0 @@
-namespace Frankie.Control
-{
-    public enum NPCMoveFocus
-    {
-        Inactive,
-        Chasing,
-        Fleeing,
-        Pending,
-        Patrolling,
-        RandomWalk
-    }
-}
