--- Assets/Scripts/Control/NPC/PatrolPath.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3511dd9d90fa09f4dbb392364e0aeb04
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
