--- Assets/Scripts/Control/NPC/PatrolPathWaypoint.cs
@@ -1,13 +1,0 @@
-using UnityEngine;
-
-namespace Frankie.Control
-{
-    public class PatrolPathWaypoint : MonoBehaviour
-    {
-        // Tunables
-        [SerializeField] WaypointType waypointType = WaypointType.Move;
-
-        // Public Methods
-        public WaypointType GetWaypointType() => waypointType;
-    }
-}
