--- Assets/Scripts/Control/NPC/WaypointType.cs
@@ -1,8 +1,0 @@
-namespace Frankie.Control
-{
-    public enum WaypointType
-    {
-        Move,
-        Warp
-    }
-}
