+++ Assets/Scripts/Zones/MoveMesh/WalkabilityGrid.cs
@@ -11,8 +11,10 @@ namespace Frankie.ZoneManagement
         public float cellSize;
         public float originX;
         public float originY;
-        public List<bool> cells = new List<bool>();
+        public List<bool> cells = new();
 
+        public bool IsEmpty() => columns == 0 || rows == 0 || cells.Count == 0;
+        
         public bool GetCell(int column, int row)
         {
             if (column < 0 || column >= columns || row < 0 || row >= rows) { return false; }
@@ -20,6 +22,6 @@ namespace Frankie.ZoneManagement
         }
 
         // Local-space cell centre (no transform applied — used via MoveMesh)
-        public Vector2 CellToLocal(int col, int row) => new Vector2(originX + (col + 0.5f) * cellSize, originY + (row + 0.5f) * cellSize);
+        public Vector2 CellToLocal(int column, int row) => new(originX + (column + 0.5f) * cellSize, originY + (row + 0.5f) * cellSize);
     }
 }
