+++ ProjectSettings/Physics2DSettings.asset
@@ -53,5 +53,5 @@ Physics2DSettings:
   m_ReuseCollisionCallbacks: 1
   m_AutoSyncTransforms: 0
   m_GizmoOptions: 10
-  m_LayerCollisionMatrix: 00e2ffff00e0ffff4ce1ffff44f0ffff00e0ffff00e0ffff4ce0ffff00e0ffff04f0ffff01e0ffff00e8ffff00e4ffff08e1ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff
+  m_LayerCollisionMatrix: 00c2ffff00c0ffff4cc1ffff44f0ffff00c0ffff00c0ffff4ce0ffff00c0ffff04f0ffff01c0ffff00c8ffff00c4ffff08c1ffff48c1ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff
   m_PhysicsLowLevelSettings: {fileID: 0}
